✏️ WOW-59 rename: fetch server, client 파일 api 폴더로 이동
