+++ src/app/api/product/route.ts
@@ -1,4 +1,4 @@
-import { fetchServer } from "@/utils/apis/fetch.server";
+import { fetchServer } from "@/api/fetch.server";
 import { NextRequest } from "next/server";
 
 export const GET = async (request: NextRequest) => {
